Removing unnecessary files
