+++ Medusa/__init__.py
@@ -1,2 +1,1 @@
-#TODO: Fill
-
+from .sort import *
